+++ CHANGELOG.md
@@ -6,4 +6,6 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- Initial release
+### Added
+
+- Provide code navigation support for `macro` and `function` in `.ftl` files.